spawn: Rework spawning
The old spawning code was copied from glib, and used a thread to
make the synchronous variant into an asynchronous one. This had the
problem that fork(3p) was called from the worker thread, not the
calling (main) thread.

The new code replaces this with a two-stage approach where the
fork takes place on the calling thread, and only the async wait
for the child error is done in a worker thread.

This also unifies the sync and async spawn variants to use
the exact same code, only called in one or two stages.

This also removes calls to setenv/unsetenv from the child setup
code, which are not safe to do there. Instead, the environment
is prepared in advance and used via execve(2).

Fixes: https://gitlab.gnome.org/GNOME/vte/-/issues/118

+++ src/vte.cc
@@ -17,7 +17,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-#include <config.h>
+#include "config.h"
 
 #include <math.h>
 #include <search.h>
@@ -10088,7 +10088,7 @@ Terminal::terminate_child() noexcept
                 return false;
 
         auto pgrp = getpgid(m_pty_pid);
-        if (pgrp != -1) {
+        if (pgrp != -1 && pgrp != getpgid(getpid())) {
                 kill(-pgrp, SIGHUP);
         }
